+++ WebApi/InWords.WebApi/InWords.Data/InWords.Data.csproj
@@ -11,8 +11,4 @@
     <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="2.1.2" />
   </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\InWords.Service.Encryption\InWords.Service.Encryption.csproj" />
-  </ItemGroup>
-
 </Project>
